+++ src/Symfony/Component/Cache/Tests/Adapter/TagAwareAdapterTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Cache\Tests\Adapter;
 
 use PHPUnit\Framework\MockObject\MockObject;
+use Psr\Cache\CacheItemInterface;
 use Psr\Cache\CacheItemPoolInterface;
 use Symfony\Component\Cache\Adapter\AdapterInterface;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
@@ -67,6 +68,42 @@ class TagAwareAdapterTest extends AdapterTestCase
         $this->assertFalse($cache->prune());
     }
 
+    public function testKnownTagVersionsTtl()
+    {
+        $itemsPool = new FilesystemAdapter('', 10);
+        $tagsPool = $this
+            ->getMockBuilder(AdapterInterface::class)
+            ->getMock();
+
+        $pool = new TagAwareAdapter($itemsPool, $tagsPool, 10);
+
+        $item = $pool->getItem('foo');
+        $item->tag(['baz']);
+        $item->expiresAfter(100);
+
+        $tag = $this->getMockBuilder(CacheItemInterface::class)->getMock();
+        $tag->expects(self::exactly(2))->method('get')->willReturn(10);
+
+        $tagsPool->expects(self::exactly(2))->method('getItems')->willReturn([
+            'baz'.TagAwareAdapter::TAGS_PREFIX => $tag,
+        ]);
+
+        $pool->save($item);
+        $this->assertTrue($pool->getItem('foo')->isHit());
+        $this->assertTrue($pool->getItem('foo')->isHit());
+
+        sleep(20);
+
+        $this->assertTrue($pool->getItem('foo')->isHit());
+
+        sleep(5);
+
+        $this->assertTrue($pool->getItem('foo')->isHit());
+    }
+
+    /**
+     * @return MockObject|PruneableCacheInterface
+     */
     private function getPruneableMock(): AdapterInterface
     {
         $pruneable = $this->createMock([PruneableInterface::class, AdapterInterface::class]);