+++ src/Symfony/Component/DependencyInjection/Loader/FileLoader.php
@@ -26,6 +26,8 @@ use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
  */
 abstract class FileLoader extends BaseFileLoader
 {
+    public const ANONYMOUS_ID_REGEXP = '/^\.\d+_[^~]++~[._a-zA-Z\d]{7}$/';
+
     protected $container;
     protected $isLoadingInstanceof = false;
     protected $instanceof = [];